+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Jump to version:
 ### MLEM
 Additions
 - Added TokenizedString.Realign
-- Added EnumHelper.GetFlags
+- Added GetFlags and GetUniqueFlags to EnumHelper
 - **Added the ability to find paths to one of multiple goals using AStar**
 
 Improvements
@@ -60,7 +60,7 @@ Fixes
 Additions
 - Added data, from, and copy instructions to DataTextureAtlas
 - Added the ability to add additional regions to a RuntimeTexturePacker after packing
-- Added DynamicEnum.GetFlags
+- Added GetFlags, GetUniqueFlags and IsDefined to DynamicEnum
 
 Improvements
 - Allow data texture atlas pivots and offsets to be negative